requirements.txt
anyio==4.8.0
arrow==1.3.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ics==0.7.2
idna==3.10
notion-client==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
requests==2.32.3
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
TatSu==5.13.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
zope.interface==7.2
